+++ locales/en/common.json
@@ -250,5 +250,6 @@
     },
     "surah-yaseen": "Surah Yaseen",
     "ayat-al-kursi": "Ayat Al-Kursi",
-    "surah-al-kahf": "Surah Al-Kahf"
+    "surah-al-kahf": "Surah Al-Kahf",
+    "surah-al-mulk": "Surah Al-Mulk"
 }
