+++ src/components/dls/Footer/Footer.tsx
@@ -1,12 +1,16 @@
+import BottomSection from './BottomSection';
 import styles from './Footer.module.scss';
 import Links from './Links';
 import TitleAndDescription from './TitleAndDescription';
 
 const Footer = () => {
   return (
-    <div className={styles.container}>
-      <TitleAndDescription />
-      <Links />
+    <div>
+      <div className={styles.container}>
+        <TitleAndDescription />
+        <Links />
+      </div>
+      <BottomSection />
     </div>
   );
 };
